refactor: rearranging components
