+++ frontend/src/components/svgs/SendSvg.tsx
@@ -22,7 +22,7 @@ function SendSvg() {
 				<path
 					d="m3.69 8.75 15.92-5.69a1 1 0 0 1 1.33 1.33l-5.69 15.92a1 1 0 0 1-1.95 0l-2.51-6.52a1 1 0 0 0-.6-.6L3.67 10.7a1 1 0 0 1 .02-1.95Z"
 					style={{
-						fill: "#2ca9bc",
+						fill: "#EAB308",
 						strokeWidth: 2,
 					}}
 				/>
